Cater for additional 'MCC 6012' options
I was recently informed that all merchants with the Merchant Category Code (MCC) of 6012 would have to add
some additional data to all card transactions. PayPoint have catered for this requirement by adding four new
option types to their messages. To avoid having to add new getters and setters for each of the four new option
types - and to avoid having to fudge the option string in the `sendData()` function - I have added a parameter
to the gateway's default parameters.
If the new parameter (`additionalOptions`) is an array, it's contents will be merged with the option array and
sent to the gateway. At this stage, no additional checks are done on the options within the array, but adding
some basic validation may be a good idea.